microservices/eth_price/requirements.txt
coinbase-advanced-py
python-dotenv
requests
web3
